% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_packs.R
\name{plot_packs}
\alias{plot_packs}
\title{Plot Packs Interactively}
\usage{
plot_packs(obs, pack)
}
\arguments{
\item{obs}{An \code{sf} object with individuals and their packs.}

\item{pack}{The name of the column in \code{obs} that contains the pack information.}
}
\value{
An interactive leaflet map.
}
\description{
Creates an interactive map visualizing the packs territories.
}
\examples{
data(samples)
samples$pack <- NA
samples[samples$Individual == "W1",]$pack <- "1"
samples[samples$Individual == "W2",]$pack <- "1"
samples[samples$Individual == "W3",]$pack <- "1"
samples[samples$Individual == "W4",]$pack <- "Lone Individual"
samples[samples$Individual == "W5",]$pack <- "2"
samples[samples$Individual == "W6",]$pack <- "3"
samples[samples$Individual == "W7",]$pack <- "2"
samples[samples$Individual == "W8",]$pack <- "2"
samples[samples$Individual == "W9",]$pack <- "3"
map <- plot_packs(samples, pack = pack)
map
}
