% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_relate.R
\docType{data}
\name{relate}
\alias{relate}
\title{relate}
\format{
A data frame with 81 rows and 3 variables:
\describe{
\item{ ind1 }{  factor }
\item{ ind2 }{  factor }
\item{ wang }{  numeric }
}
}
\source{
Source
}
\usage{
relate
}
\description{
Description.
}
\keyword{datasets}
