% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_samples.R
\docType{data}
\name{samples}
\alias{samples}
\title{samples}
\format{
A data frame with 49 rows and 3 variables:
\describe{
\item{ Individual }{  character }
\item{ geometry }{  sfc_POINT,sfc }
\item{ Sex }{  character }
}
}
\source{
Source
}
\usage{
samples
}
\description{
Description.
}
\keyword{datasets}
