% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{WolfPackR}
\alias{WolfPackR-package}
\alias{WolfPackR}
\title{WolfPackR: Analyzing Wolf Pack Territories}
\description{
A comprehensive toolkit for analyzing genetic and spatial structures of wolf packs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eboncourt/WolfPackR}
  \item Report bugs at \url{https://github.com/eboncourt/WolfPackR/issues}
}

}
\author{
\strong{Maintainer}: Etienne Boncourt \email{etienne.boncourt@gmail.com} (\href{https://orcid.org/0000-0001-8233-0190}{ORCID})

}
