% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_sf.R
\name{mcp_sf}
\alias{mcp_sf}
\title{Calculate Minimum Convex Polygon (MCP) with Percentile Filtering}
\usage{
mcp_sf(data, percentile = 100, buffer_radius = 0.01)
}
\arguments{
\item{data}{An \code{sf} object or a data.frame with columns \code{X} and \code{Y}.}

\item{percentile}{The percentile of points (distance from centroid) excluded before calculating the MCP. Default is 100.}

\item{buffer_radius}{The buffer distance around the coordinates if there is insufficient data to calculate a MCP (less than 3 data points) (default: 0.01).}
}
\value{
An \code{sf} object representing the MCP, or \code{NULL} if insufficient points.
}
\description{
Computes the MCP for a set of spatial points using \code{sf}, with an option to exclude points beyond a given percentile of distance from the centroid. Largely inspired by juoe (https://rdrr.io/github/juoe/sdmflow/).
}
\examples{
obs <- data.frame(X = c(1,2,2,5,5,6,7,8,9), Y = c(1,2,9,4,5,6,1,8,9))
obs_sf <- st_as_sf(obs, coords = c("X", "Y"))
mcp_sf(obs_sf, percentile=100, buffer_radius=0.01)
}
