% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_sf.R
\name{mcp_sf}
\alias{mcp_sf}
\title{Calculate Minimum Convex Polygon (MCP) with Percentile Filtering}
\usage{
mcp_sf(data, percentile = 100)
}
\arguments{
\item{data}{An \code{sf} object or a data.frame with columns \code{X} and \code{Y}.}

\item{percentile}{The percentile of points (distance from centroid) excluded before calculating the MCP. Default is 100.}
}
\value{
An \code{sf} object representing the MCP, or \code{NULL} if insufficient points.
}
\description{
Computes the MCP for a set of spatial points using \code{sf}, with an option to exclude points beyond a given percentile of distance from the centroid. Largely inspired by juoe (https://rdrr.io/github/juoe/sdmflow/).
}
\examples{
obs_i <- subset(obs, group == 1)
mcp_i <- mcp_sf(obs_i, percentile = 95)
}
