% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_packs.R
\name{plot_packs}
\alias{plot_packs}
\title{Plot Packs Interactively}
\usage{
plot_packs(obs, ud_df)
}
\arguments{
\item{obs}{An \code{sf} object with individuals and their genetic groups.}

\item{ud_df}{A data.frame containing individuals, their genetic group, and their assigned pack.}
}
\value{
An interactive leaflet map.
}
\description{
Creates an interactive map visualizing the packs territories.
}
\examples{
map <- plot_groups(obs, ud_df = ud_df)
map
}
