% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_packs.R
\name{plot_packs}
\alias{plot_packs}
\title{Plot Packs Interactively}
\usage{
plot_packs(obs, pack)
}
\arguments{
  \item{obs}{An `sf` object with individuals and their packs.}
  \item{pack}{The name of the column in `obs` that contains the pack information.}
}
\value{
An interactive leaflet map.
}
\description{
Creates an interactive map visualizing the packs territories.
}
\examples{
# Example usage:
map <- plot_packs(obs, pack = Pack)
map
}
