% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_samples_Carpathians.R
\docType{data}
\name{samples_Carpathians}
\alias{samples_Carpathians}
\title{samples_Carpathians}
\format{
A data frame with 277  rows and 3 variables:
\describe{
\item{ Individual }{  character }
\item{ Sex }{  character }
\item{ geometry }{  sfc_POINT,sfc }
}
}
\source{
Source
}
\usage{
samples_Carpathians
}
\description{
Description.
An sf object containing individual names, sex and coordinates associated with 277 samples collected in Romanian Carpathians. Source: Foundation Conservation Carpathia. (2024). Wolf population size and composition in one of Europe’s strongholds, the Romanian Carpathians \link{Datasat}. Zenodo. https://doi.org/10.5281/ZENODO.14544266
}
\keyword{datasets}
