% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_packs.R
\name{summarise_packs}
\alias{summarise_packs}
\title{Summarise Packs and Individuals}
\usage{
summarise_packs(obs, pack, sex_column = "Sex", male_pattern = "M", female_pattern = "F")
}
\arguments{
  \item{obs}{An `sf` object with individuals and their packs.}
  \item{pack}{The name of the column in `obs` that contains the pack information.}
  \item{sex_column}{The name of the column in `obs` that contains sex information (default: "Sex").}
  \item{male_pattern}{A regex pattern to identify males in Individual IDs (default: "M").}
  \item{female_pattern}{A regex pattern to identify females in Individual IDs (default: "F").}
}
\value{
A list containing detailed summaries of packs and individuals.
}
\description{
Summarises individuals and their packs, including putative dominant individuals.
}
\examples{
# Example usage:
summary <- summarise_packs(obs, pack = Pack)
}
