% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_packs.R
\name{summarise_packs}
\alias{summarise_packs}
\title{Summarise Packs and Individuals}
\usage{
summarise_packs(
  obs,
  pack,
  sex_column = "Sex",
  male_pattern = "M",
  female_pattern = "F"
)
}
\arguments{
\item{obs}{An \code{sf} object with individuals and their packs.}

\item{pack}{The name of the column in \code{obs} that contains the pack information.}

\item{sex_column}{The name of the column in obs that contains sex information (default: "Sex").}

\item{male_pattern}{A regex pattern to identify males in Individual IDs (default: "M").}

\item{female_pattern}{A regex pattern to identify females in Individual IDs (default: "F").}
}
\value{
A list containing detailed summaries of packs and individuals.
}
\description{
Summarises individuals and their packs, including putative dominant individuals.
}
\examples{
data(samples)
samples$pack <- NA
samples[samples$Individual == "W1",]$pack <- "1"
samples[samples$Individual == "W2",]$pack <- "1"
samples[samples$Individual == "W3",]$pack <- "1"
samples[samples$Individual == "W4",]$pack <- "Lone Individual"
samples[samples$Individual == "W5",]$pack <- "2"
samples[samples$Individual == "W6",]$pack <- "3"
samples[samples$Individual == "W7",]$pack <- "2"
samples[samples$Individual == "W8",]$pack <- "2"
samples[samples$Individual == "W9",]$pack <- "3"
samples[samples$Individual == "W1",]$Individual <- "M1"
samples[samples$Individual == "W2",]$Individual <- "M2"
samples[samples$Individual == "W3",]$Individual <- "M3"
samples[samples$Individual == "W4",]$Individual <- "M4"
samples[samples$Individual == "W5",]$Individual <- "F1"
samples[samples$Individual == "W6",]$Individual <- "F2"
samples[samples$Individual == "W7",]$Individual <- "F3"
samples[samples$Individual == "W8",]$Individual <- "F4"
samples[samples$Individual == "W9",]$Individual <- "F5"
summary <- summarise_packs(samples, pack, male_pattern = "M", female_pattern = "F")
}
